Move RetractArms action to impl module
